+++ src/attrs.cpp
@@ -98,3 +98,7 @@ void cpp_copy_names(SEXP source, SEXP target, bool deep_copy){
 void cpp_shallow_duplicate_attrs(SEXP source, SEXP target){
   SHALLOW_DUPLICATE_ATTRIB(target, source);
 }
+[[cpp11::register]]
+void cpp_copy_most_attrs(SEXP source, SEXP target){
+  Rf_copyMostAttrib(source, target);
+}