+++ src/cpp11.cpp
@@ -55,6 +55,14 @@ extern "C" SEXP _cheapr_cpp_shallow_duplicate_attrs(SEXP source, SEXP target) {
     return R_NilValue;
   END_CPP11
 }
+// attrs.cpp
+void cpp_copy_most_attrs(SEXP source, SEXP target);
+extern "C" SEXP _cheapr_cpp_copy_most_attrs(SEXP source, SEXP target) {
+  BEGIN_CPP11
+    cpp_copy_most_attrs(cpp11::as_cpp<cpp11::decay_t<SEXP>>(source), cpp11::as_cpp<cpp11::decay_t<SEXP>>(target));
+    return R_NilValue;
+  END_CPP11
+}
 // gcd.cpp
 double cpp_gcd2(double x, double y, double tol, bool na_rm);
 extern "C" SEXP _cheapr_cpp_gcd2(SEXP x, SEXP y, SEXP tol, SEXP na_rm) {
@@ -469,10 +477,10 @@ extern "C" SEXP _cheapr_cpp_sset_range(SEXP x, SEXP from, SEXP to, SEXP by) {
   END_CPP11
 }
 // sset.cpp
-SEXP cpp_df_select(SEXP x, SEXP locs, bool copy_attrs);
-extern "C" SEXP _cheapr_cpp_df_select(SEXP x, SEXP locs, SEXP copy_attrs) {
+SEXP cpp_df_select(SEXP x, SEXP locs);
+extern "C" SEXP _cheapr_cpp_df_select(SEXP x, SEXP locs) {
   BEGIN_CPP11
-    return cpp11::as_sexp(cpp_df_select(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(locs), cpp11::as_cpp<cpp11::decay_t<bool>>(copy_attrs)));
+    return cpp11::as_sexp(cpp_df_select(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(locs)));
   END_CPP11
 }
 // sset.cpp
@@ -483,10 +491,17 @@ extern "C" SEXP _cheapr_cpp_sset_df(SEXP x, SEXP indices) {
   END_CPP11
 }
 // sset.cpp
-SEXP cpp_df_subset(SEXP x, SEXP i, SEXP j);
-extern "C" SEXP _cheapr_cpp_df_subset(SEXP x, SEXP i, SEXP j) {
+SEXP cpp_df_slice(SEXP x, SEXP indices);
+extern "C" SEXP _cheapr_cpp_df_slice(SEXP x, SEXP indices) {
   BEGIN_CPP11
-    return cpp11::as_sexp(cpp_df_subset(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(i), cpp11::as_cpp<cpp11::decay_t<SEXP>>(j)));
+    return cpp11::as_sexp(cpp_df_slice(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(indices)));
+  END_CPP11
+}
+// sset.cpp
+SEXP cpp_df_subset(SEXP x, SEXP i, SEXP j, bool keep_attrs);
+extern "C" SEXP _cheapr_cpp_df_subset(SEXP x, SEXP i, SEXP j, SEXP keep_attrs) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_df_subset(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(i), cpp11::as_cpp<cpp11::decay_t<SEXP>>(j), cpp11::as_cpp<cpp11::decay_t<bool>>(keep_attrs)));
   END_CPP11
 }
 // utils.cpp
@@ -605,12 +620,14 @@ static const R_CallMethodDef CallEntries[] = {
     {"_cheapr_cpp_col_all_na",              (DL_FUNC) &_cheapr_cpp_col_all_na,              2},
     {"_cheapr_cpp_col_any_na",              (DL_FUNC) &_cheapr_cpp_col_any_na,              2},
     {"_cheapr_cpp_col_na_counts",           (DL_FUNC) &_cheapr_cpp_col_na_counts,           2},
+    {"_cheapr_cpp_copy_most_attrs",         (DL_FUNC) &_cheapr_cpp_copy_most_attrs,         2},
     {"_cheapr_cpp_count_val",               (DL_FUNC) &_cheapr_cpp_count_val,               3},
     {"_cheapr_cpp_dbl_sequence",            (DL_FUNC) &_cheapr_cpp_dbl_sequence,            3},
     {"_cheapr_cpp_df_col_na_counts",        (DL_FUNC) &_cheapr_cpp_df_col_na_counts,        1},
     {"_cheapr_cpp_df_row_na_counts",        (DL_FUNC) &_cheapr_cpp_df_row_na_counts,        1},
-    {"_cheapr_cpp_df_select",               (DL_FUNC) &_cheapr_cpp_df_select,               3},
-    {"_cheapr_cpp_df_subset",               (DL_FUNC) &_cheapr_cpp_df_subset,               3},
+    {"_cheapr_cpp_df_select",               (DL_FUNC) &_cheapr_cpp_df_select,               2},
+    {"_cheapr_cpp_df_slice",                (DL_FUNC) &_cheapr_cpp_df_slice,                2},
+    {"_cheapr_cpp_df_subset",               (DL_FUNC) &_cheapr_cpp_df_subset,               4},
     {"_cheapr_cpp_drop_null",               (DL_FUNC) &_cheapr_cpp_drop_null,               2},
     {"_cheapr_cpp_fixed_width_breaks",      (DL_FUNC) &_cheapr_cpp_fixed_width_breaks,      6},
     {"_cheapr_cpp_format_numeric_as_int64", (DL_FUNC) &_cheapr_cpp_format_numeric_as_int64, 1},